Refactor blocks/partials
Now that I understand blocks and partials, and now that I understand the
layout and look of my site better, pull a bunch more stuff into baseof.

The headers were basically all the same, and it turns out I don't need
to define blocks I don't need on the page.